move concat to common.go
