+++ common/common.go
@@ -1,0 +1,9 @@
+package common
+
+// Concat concatenates two byte arrays
+func Concat(s1 []byte, s2 ...byte) []byte {
+	r := make([]byte, len(s1)+len(s2))
+	copy(r, s1)
+	copy(r[len(s1):], s2)
+	return r
+}
